More player functions
Add turn() function for player
Add buy_card() for player
Add skeleton function to count player's victory points
Use i32 everywhere

+++ src/game/cards/dominion.rs
@@ -2,7 +2,7 @@ use crate::game::{player::Player, traits::*};
 
 pub struct Smithy;
 impl Card for Smithy {
-    fn cost(&self) -> u8 {
+    fn cost(&self) -> i32 {
         return 4;
     }
